requirements.txt
chromadb==0.4.22
sentence-transformers==2.2.2
google-cloud-aiplatform==1.38.1
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
